create util functions for handling error and is empty response
