+++ src/controllers/userController.ts
@@ -8,6 +8,8 @@ import { MESSAGES } from '@utils/message';
 import { IUser } from '@models/User';
 import { sendResponse } from '@utils/Respons/response';
 import { AuthRequest } from 'types/auth-request';
+import { handleEmptyResponse } from '@utils/Respons/handleEmptyResponse';
+import handleErrorResponse from '@utils/Respons/handleErrorResponse';
 
 export const createUser = async (req: Request, res: Response) => {
   try {
@@ -20,19 +22,14 @@ export const createUser = async (req: Request, res: Response) => {
     const user : IUser | null = await userService.createUser(req.body);
     sendResponse(res, HTTP_STATUS.CREATED, MESSAGES.SIGNUP_SUCCESSFULLY, user);
   } catch (error) {
-    if (error instanceof mongoose.Error.ValidationError) {
-      const errors = Object.values(error.errors).map((err) => err.message);
-      return sendResponse(res, HTTP_STATUS.BAD_REQUEST, errors);
-    } else {
-      return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, MESSAGES.INTERNAL_SERVER_ERROR);
-    }
+    handleErrorResponse(res, error);
   }
 };
 
 export const getUsers = async (req: AuthRequest, res: Response) => {
   try {    
     const users : IUser[]= await userService.getUsers();
-    sendResponse(res, HTTP_STATUS.OK, MESSAGES.RECORD_FETCHED_SUCCESSFULLY, users);
+    handleEmptyResponse(res, users);
   } catch (error) {
     sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, MESSAGES.INTERNAL_SERVER_ERROR);
   }
@@ -41,11 +38,7 @@ export const getUsers = async (req: AuthRequest, res: Response) => {
 export const getUserById = async (req: AuthRequest, res: Response) => {
   try {
     const user : IUser | null = await userService.getUserById(req.params.userId);
-    if (user) {
-      sendResponse(res, HTTP_STATUS.OK, MESSAGES.RECORD_FETCHED_SUCCESSFULLY, user);
-    } else {
-      sendResponse(res, HTTP_STATUS.NOT_FOUND, MESSAGES.NO_RECORD);
-    }
+    handleEmptyResponse(res, user);
   } catch (error) {
     sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, MESSAGES.INTERNAL_SERVER_ERROR);
   }
@@ -67,18 +60,9 @@ export const updateUser = async (req: Request, res: Response) => {
       }
     }    
     const user = await userService.updateUser(userId, userData);
-    if (user) {
-      sendResponse(res, HTTP_STATUS.OK, MESSAGES.RECORD_UPDATED_SUCCESSFULLY, user);
-    } else {
-      sendResponse(res, HTTP_STATUS.NOT_FOUND, MESSAGES.NO_RECORD);
-    }
+    handleEmptyResponse(res, user);
   } catch (error) {
-    if (error instanceof mongoose.Error.ValidationError) {
-      const errors = Object.values(error.errors).map((err) => err.message);
-      return sendResponse(res, HTTP_STATUS.BAD_REQUEST, errors);
-    } else {
-      return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, MESSAGES.INTERNAL_SERVER_ERROR);
-    }
+    handleErrorResponse(res, error);
   }
 };
